# packages/reservation/openapi.yaml
openapi: 3.0.1
info:
  title: Reservation API
  version: 1.0.0
components:
  securitySchemes:
    bearerAuthJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TestResult200:
      type: object
      deprecated: false
      additionalProperties: false
      properties:
        id:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        testType:
          type: string
        dateOfAppointment:
          type: string
        result:
          type: string
...
    PCRTestResultsResponse200:
      type: object
      additionalProperties: false
      properties:
        data:
          type: object
          additionalProperties: false
        status:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - succeed
                - failed
                - validation_error
                - internal_server_error
                - unauthorized
                - access_denied
                - resource_not_found
            message:
              nullable: true
              type: string
...
    TestResultsRequestBulk:
      type: object
      additionalProperties: false
      properties:
        reportTrackerId:
          type: string
        resultDate:
          type: string
        templateId:
          type: string
        labId:
          type: string
        fileName:
          type: string
        results:
          type: array
          items:
            type: object
            required:
              - action
              - autoResult
              - barCode
              - resultAnalysis
            properties:
              action:
                type: string
                default: SendThisResult
                enum:
                  - SendThisResult
                  - DoNothing
                  - ReRunToday
                  - ReRunTomorrow
                  - RequestReCollect
                  - RecollectAsInvalid
                  - RecollectAsInconclusive
                  - MarkAsPositive
                  - MarkAsPresumptivePositive
                  - MarkAsNegative
                  - SendPreliminaryPositive
              autoResult:
                type: string
                enum:
                  - PresumptivePositive
                  - Positive
                  - Negative
                  - Invalid
                  - Inconclusive
              barCode:
                type: string
                minLength: 1
              notify:
                type: boolean
                default: false
              resultAnalysis:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  properties:
                    label:
                      type: string
                      enum:
                        - famEGene
                        - famCt
                        - calRedRdrp
                        - calRedCt
                        - quasarNgene
                        - quasarCt
                        - hexIc
                        - hexCt
                        - ORF1abCt
                        - SGeneCt
                        - NGeneCt
                        - MS2Ct
                        - ORF1ab
                        - SGene
                        - NGene
                        - MS2
                        - IgAResult
                        - IgGResult
                        - IgMResult
                        - IgA
                        - IgG
                        - IgM
                    value:
                      type: string
              comment:
                type: string
      required:
        - results
        - resultDate
        - templateId
        - labId
        - fileName
...
paths:
...
  /reservation/admin/api/v1/test-results:
    get:
      description: Get Test results
      tags:
        - Test Management
      parameters:
        - in: query
          name: organizationId
          schema:
            type: string
        - in: query
          name: dateOfAppointment
          schema:
            type: string
      responses:
        '200':
          description: Successfully Return
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestResult200'
...
  /reservation/admin/api/v1/test-results-bulk:
    post:
      description: Send Multiple PCR Results
      tags:
        - Test Results Management
      security:
        - bearerAuthJWT: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestResultsRequestBulk'
      responses:
        '200':
          description: Successfully Return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PCRTestResultsResponse200'
  /reservation/admin/api/v1/pcr-test-results/list/stats:
    get:
      description: Retrieve list of PCR Results
      tags:
        - Test Results Management
      security:
        - bearerAuthJWT: []
      parameters:
        - in: query
          name: labId
          schema:
            type: string
        - in: query
          name: organizationId
          schema:
            type: string
        - in: query
          name: date
          schema:
            type: string
            format: date
        - in: query
          name: barCode
          schema:
            type: string
        - in: query
          name: testType
          schema:
            type: string
        - in: query
          name: searchQuery
          schema:
            type: string
        - in: query
          name: result
          schema:
            type: string
            enum:
              - PresumptivePositive
              - PreliminaryPositive
              - Positive
              - Negative
              - Pending
              - Invalid
              - Inconclusive
              - Indeterminate
      responses:
        '200':
          description: Succesfully return stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse200'
  /reservation/admin/api/v1/pcr-test-results:
    post:
      description: Send Test Results
      tags:
        - Test Results Management
      security:
        - bearerAuthJWT: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestResultsRequest'
      responses:
        '200':
          description: Successfully Return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PCRTestResultsResponse200'
    get:
      description: Retrieve list of PCR Results
      tags:
        - Test Results Management
      security:
        - bearerAuthJWT: []
      parameters:
        - in: query
          name: labId
          schema:
            type: string
        - in: query
          name: organizationId
          schema:
            type: string
        - in: query
          name: date
          schema:
            type: string
            format: date
        - in: query
          name: barCode
          schema:
            type: string
        - in: query
          name: testType
          schema:
            type: string
            enum:
              - PCR
              - RapidAntigen
        - in: query
          name: result
          schema:
            type: string
            enum:
              - PresumptivePositive
              - PreliminaryPositive
              - Positive
              - Negative
              - Pending
              - Invalid
              - Inconclusive
              - Indeterminate
        - in: query
          name: searchQuery
          schema:
            type: string
      responses:
        '200':
          description: Successfully Return
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      properties:
                        id:
                          type: string
                        firstName:
                          type: string
                        lastName:
                          type: string
                        testType:
                          type: string
                        dateOfAppointment:
                          type: string
                        barCode:
                          type: string
                        result:
                          type: string
                          enum:
                            - PresumptivePositive
                            - Positive
                            - Negative
                            - Pending
                            - Invalid
                            - Inconclusive
                            - PreliminaryPositive
                        dateTime:
                          type: string
                        deadline:
                          type: string
                        previousResult:
                          type: string
                          enum:
                            - PresumptivePositive
                            - Positive
                            - Negative
                            - Pending
                            - Invalid
                            - Inconclusive
                            - PreliminaryPositive
                        organizationName:
                          type: string
                        style:
                          type: string
                          enum:
                            - RED
                            - GREEN
                            - YELLOW
                            - BLUE
                            - GREY
                        testName:
                          type: string
                        doctorId:
                          type: string
                        appointmentStatus:
                          type: string
                          enum:
                            - Pending
                            - Submitted
                            - InTransit
                            - Received
                            - CheckedIn
                            - InProgress
                            - Reported
                            - ReRunRequired
                            - ReCollectRequired
                            - Canceled
                        labName:
                          type: string
                  status:
                    type: object
                    additionalProperties: false
                    properties:
                      code:
                        type: string
                        enum:
                          - succeed
                          - failed
                          - validation_error
                          - internal_server_error
                          - unauthorized
                          - access_denied
                          - resource_not_found
                      message:
                        nullable: true
                        type: string
                  page:
                    type: number
  /reservation/admin/api/v1/pcr-test-results/confirm:
    post:
      description: Confirm PCR Results
      tags:
        - Test Results Management
      security:
        - bearerAuthJWT: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                barCode:
                  type: string
                action:
                  type: string
                  enum:
                    - Indeterminate
                    - MarkAsPositive
                    - MarkAsNegative
              required:
                - barCode
                - action
      responses:
        '200':
          description: Succesfully add label
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentTransportRunResponse200'
...
  /reservation/admin/api/v1/pcr-test-results-bulk/report-status:
    get:
      description: Get appointment by barCode
      tags:
        - Appointment Management
      security:
        - bearerAuthJWT: []
      parameters:
        - in: query
          name: reportTrackerId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successfully Return
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      properties:
                        details:
                          type: string
                        barCode:
                          type: string
                        status:
                          type: string
                          enum:
                            - Failed
                            - Processing
                            - RequestReceived
                            - SentReRunRequest
                            - SentReCollectRequest
                            - SentResult
                            - Skipped
                      required:
                        - barCode
                        - status
                  status:
                    type: object
                    additionalProperties: false
                    properties:
                      code:
                        type: string
                        enum:
                          - succeed
                          - failed
                          - validation_error
                          - internal_server_error
                          - unauthorized
                          - access_denied
                          - resource_not_found
                      message:
                        nullable: true
                        type: string
                  page:
                    type: number
  /reservation/admin/api/v1/pcr-test-results/add-test-run:
    put:
      description: Add transport run to appointment
      tags:
        - Transport Runs
      security:
        - bearerAuthJWT: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                pcrTestResultIds:
                  type: array
                  items:
                    type: string
                testRunId:
                  type: string
              required:
                - pcrTestResultIds
                - testRunId
      responses:
        '200':
          description: Succesfully added PcrTestResult
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      properties:
                        id:
                          type: string
                        reason:
                          type: string
                        barCode:
                          type: string
                        status:
                          type: string
                          enum:
                            - Success
                            - Failed
                      required:
                        - id
                        - status
                  status:
                    type: object
                    additionalProperties: false
                    properties:
                      code:
                        type: string
                        enum:
                          - succeed
                          - failed
                          - validation_error
                          - internal_server_error
                          - unauthorized
                          - access_denied
                          - resource_not_found
                      message:
                        nullable: true
                        type: string
                  page:
                    type: number
  /reservation/admin/api/v1/pcr-test-results/history:
    post:
      description: Get PCR Test Results by Barcode
      security:
        - bearerAuthJWT: []
      tags:
        - Test Results Management
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                barcode:
                  type: array
                  items:
                    type: string
              required:
                - barcode
      responses:
        '200':
          description: Successfully Return
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      properties:
                        id:
                          type: string
                        barCode:
                          type: string
                        waitingResult:
                          type: boolean
                        results:
                          type: array
                          items:
                            type: object
                            additionalProperties: false
                            properties:
                              resultAnalysis:
                                type: array
                                items:
                                  type: object
                                  additionalProperties: false
                                  properties:
                                    label:
                                      type: string
                                    value:
                                      type: string
                              result:
                                type: string
                        reason:
                          type: string
                          enum:
                            - Already Reported
                            - ReCollect Already Requested
                            - In Progress
                            - No In Progress
                      required:
                        - id
                        - firstName
                        - lastName
                        - barCode
                  status:
                    type: object
                    additionalProperties: false
                    properties:
                      code:
                        type: string
                        enum:
                          - succeed
                          - failed
                          - validation_error
                          - internal_server_error
                          - unauthorized
                          - access_denied
                          - resource_not_found
                      message:
                        nullable: true
                        type: string
                  page:
                    type: number
...
  /reservation/api/v1/test-results/{id}:
    get:
      description: Details for PCR Results for Regular User
      tags:
        - Test Results Management
      security:
        - bearerAuthJWT: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successfully Return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResultDetailResponse'
...
  /reservation/admin/api/v1/test-results/{id}:
    get:
      description: Details for PCR Results for Admin
      tags:
        - Test Results Management
      security:
        - bearerAuthJWT: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successfully Return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResultDetailResponse'
...
  /reservation/admin/api/v1/pcr-test-results/due-deadline/list/stats:
    get:
      description: Retrieve list of PCR Results
      tags:
        - Test Results Management
      security:
        - bearerAuthJWT: []
      parameters:
        - in: query
          name: labId
          schema:
            type: string
        - in: query
          name: deadline
          schema:
            type: string
            format: date
        - in: query
          name: testRunId
          schema:
            type: string
        - in: query
          name: barCode
          schema:
            type: string
      responses:
        '200':
          description: Succesfully return stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse200'
...